新しくMountainsモデル、マイグレーションの作成、seeder名の変更
